move some more stuff, fix imports
