+++ retrieval.py
@@ -2,6 +2,9 @@ import re
 import asyncio
 import warnings
 import logging
+from typing import List, Dict, Iterable, Callable, Iterator
+from collections import defaultdict
+from itertools import chain
 
 import aiohttp
 import requests
@@ -12,25 +15,23 @@ import optimum.bettertransformer.transformation
 from sentence_transformers import SentenceTransformer
 
 
-import pickle
-
 import faiss
 import numpy as np
 
-from utils import Document, weighted_reciprocal_rank, cosine_similarity, filter_similar_embeddings
-from retrievers.bm25_retriever import BM25Retriever
-from chunkers.character_chunker import RecursiveCharacterTextSplitter
+
 try:
     from qdrant_client import QdrantClient, models
 except ImportError:
     qrant_client = None
 
 try:
-    from retrievers.qdrant_retriever import MyQdrantSparseVectorRetriever
-    from chunkers.semantic_chunker import BoundedSemanticChunker
+    from .retrievers.qdrant_retriever import MyQdrantSparseVectorRetriever
+    from .chunkers.semantic_chunker import BoundedSemanticChunker
+    from .utils import Document, cosine_similarity
 except ImportError:
     from retrievers.qdrant_retriever import MyQdrantSparseVectorRetriever
     from chunkers.semantic_chunker import BoundedSemanticChunker
+    from utils import Document, cosine_similarity
 
 
 class DocumentRetriever:
@@ -81,10 +82,8 @@ class DocumentRetriever:
 
     def retrieve_documents(self, query: str, url_list: list[str]) -> list[Document]:
         yield "Downloading webpages..."
-        #html_url_tupls = zip(asyncio.run(async_fetch_urls(url_list)), url_list)
-        #html_url_tupls = [(content, url) for content, url in html_url_tupls if content is not None]
-        with open("obama_html_tups", "rb") as f:
-            html_url_tupls = pickle.load(f)
+        html_url_tupls = zip(asyncio.run(async_fetch_urls(url_list)), url_list)
+        html_url_tupls = [(content, url) for content, url in html_url_tupls if content is not None]
         if not html_url_tupls:
             return []
 
@@ -144,7 +143,7 @@ class DocumentRetriever:
         else:
             raise ValueError("self.keyword_retriever must be one of ('bm25', 'splade')")
 
-        query_embedding = self.embeddings.encode(query)
+        query_embedding = self.embedding_model.encode(query)
         D, I = faiss_index.search(query_embedding.reshape(1, -1), self.num_results)
         result_indices = I[0]
         relevant_doc_embeddings = document_embeddings[result_indices]
@@ -223,3 +222,76 @@ def html_to_plaintext_doc(html_text: str or bytes, url: str) -> Document:
     strings = '\n'.join([s.strip() for s in soup.stripped_strings])
     webpage_document = Document(page_content=strings, metadata={"source": url})
     return webpage_document
+
+
+def weighted_reciprocal_rank(doc_lists: List[List[Document]], weights: List[float], c: int = 60) -> List[Document]:
+    """
+    Perform weighted Reciprocal Rank Fusion on multiple rank lists.
+    You can find more details about RRF here:
+    https://plg.uwaterloo.ca/~gvcormac/cormacksigir09-rrf.pdf
+
+    Args:
+        doc_lists: A list of rank lists, where each rank list contains unique items.
+        weights: A list of weights corresponding to the rank lists. Defaults to equal
+            weighting for all lists.
+        c: A constant added to the rank, controlling the balance between the importance
+            of high-ranked items and the consideration given to lower-ranked items.
+            Default is 60.
+
+    Returns:
+        list: The final aggregated list of items sorted by their weighted RRF
+                scores in descending order.
+    """
+    if len(doc_lists) != len(weights):
+        raise ValueError(
+            "Number of rank lists must be equal to the number of weights."
+        )
+
+    # Associate each doc's content with its RRF score for later sorting by it
+    # Duplicated contents across retrievers are collapsed & scored cumulatively
+    rrf_score: Dict[str, float] = defaultdict(float)
+    for doc_list, weight in zip(doc_lists, weights):
+        for rank, doc in enumerate(doc_list, start=1):
+            rrf_score[doc.page_content] += weight / (rank + c)
+
+    # Docs are deduplicated by their contents then sorted by their scores
+    all_docs = chain.from_iterable(doc_lists)
+    sorted_docs = sorted(
+        unique_by_key(all_docs, lambda doc: doc.page_content),
+        reverse=True,
+        key=lambda doc: rrf_score[doc.page_content],
+    )
+    return sorted_docs
+
+
+def filter_similar_embeddings(
+    embedded_documents: List[List[float]], similarity_fn: Callable, threshold: float
+) -> List[int]:
+    """Filter redundant documents based on the similarity of their embeddings."""
+    similarity = np.tril(similarity_fn(embedded_documents, embedded_documents), k=-1)
+    redundant = np.where(similarity > threshold)
+    redundant_stacked = np.column_stack(redundant)
+    redundant_sorted = np.argsort(similarity[redundant])[::-1]
+    included_idxs = set(range(len(embedded_documents)))
+    for first_idx, second_idx in redundant_stacked[redundant_sorted]:
+        if first_idx in included_idxs and second_idx in included_idxs:
+            # Default to dropping the second document of any highly similar pair.
+            included_idxs.remove(second_idx)
+    return list(sorted(included_idxs))
+
+
+def unique_by_key(iterable: Iterable, key: Callable) -> Iterator:
+    """Yield unique elements of an iterable based on a key function.
+
+    Args:
+        iterable: The iterable to filter.
+        key: A function that returns a hashable key for each element.
+
+    Yields:
+        Unique elements of the iterable based on the key function.
+    """
+    seen = set()
+    for e in iterable:
+        if (k := key(e)) not in seen:
+            seen.add(k)
+            yield e
