+++ utils.py
@@ -1,6 +1,3 @@
-from typing import List, Dict, Iterable, Callable, Iterator
-from collections import defaultdict
-from itertools import chain
 from dataclasses import dataclass
 
 import numpy as np
@@ -10,6 +7,17 @@ import numpy as np
 class Document:
     page_content: str
     metadata: str
+
+
+class Generator:
+    """Allows a generator method to return a final value after finishing
+    the generation. Credit: https://stackoverflow.com/a/34073559"""
+    def __init__(self, gen):
+        self.gen = gen
+
+    def __iter__(self):
+        self.retval = yield from self.gen
+        return self.retval
 
 
 def cosine_similarity(X, Y) -> np.ndarray:
@@ -33,74 +41,16 @@ def cosine_similarity(X, Y) -> np.ndarray:
     return similarity
 
 
-def weighted_reciprocal_rank(doc_lists: List[List[Document]], weights: List[float], c: int = 60) -> List[Document]:
-    """
-    Perform weighted Reciprocal Rank Fusion on multiple rank lists.
-    You can find more details about RRF here:
-    https://plg.uwaterloo.ca/~gvcormac/cormacksigir09-rrf.pdf
-
-    Args:
-        doc_lists: A list of rank lists, where each rank list contains unique items.
-        weights: A list of weights corresponding to the rank lists. Defaults to equal
-            weighting for all lists.
-        c: A constant added to the rank, controlling the balance between the importance
-            of high-ranked items and the consideration given to lower-ranked items.
-            Default is 60.
-
-    Returns:
-        list: The final aggregated list of items sorted by their weighted RRF
-                scores in descending order.
-    """
-    if len(doc_lists) != len(weights):
-        raise ValueError(
-            "Number of rank lists must be equal to the number of weights."
-        )
-
-    # Associate each doc's content with its RRF score for later sorting by it
-    # Duplicated contents across retrievers are collapsed & scored cumulatively
-    rrf_score: Dict[str, float] = defaultdict(float)
-    for doc_list, weight in zip(doc_lists, weights):
-        for rank, doc in enumerate(doc_list, start=1):
-            rrf_score[doc.page_content] += weight / (rank + c)
-
-    # Docs are deduplicated by their contents then sorted by their scores
-    all_docs = chain.from_iterable(doc_lists)
-    sorted_docs = sorted(
-        unique_by_key(all_docs, lambda doc: doc.page_content),
-        reverse=True,
-        key=lambda doc: rrf_score[doc.page_content],
-    )
-    return sorted_docs
-
-
-def filter_similar_embeddings(
-    embedded_documents: List[List[float]], similarity_fn: Callable, threshold: float
-) -> List[int]:
-    """Filter redundant documents based on the similarity of their embeddings."""
-    similarity = np.tril(similarity_fn(embedded_documents, embedded_documents), k=-1)
-    redundant = np.where(similarity > threshold)
-    redundant_stacked = np.column_stack(redundant)
-    redundant_sorted = np.argsort(similarity[redundant])[::-1]
-    included_idxs = set(range(len(embedded_documents)))
-    for first_idx, second_idx in redundant_stacked[redundant_sorted]:
-        if first_idx in included_idxs and second_idx in included_idxs:
-            # Default to dropping the second document of any highly similar pair.
-            included_idxs.remove(second_idx)
-    return list(sorted(included_idxs))
-
-
-def unique_by_key(iterable: Iterable, key: Callable) -> Iterator:
-    """Yield unique elements of an iterable based on a key function.
-
-    Args:
-        iterable: The iterable to filter.
-        key: A function that returns a hashable key for each element.
-
-    Yields:
-        Unique elements of the iterable based on the key function.
-    """
-    seen = set()
-    for e in iterable:
-        if (k := key(e)) not in seen:
-            seen.add(k)
-            yield e
+def dict_list_to_pretty_str(data: list[dict]) -> str:
+    ret_str = ""
+    if isinstance(data, dict):
+        data = [data]
+    if isinstance(data, list):
+        for i, d in enumerate(data):
+            ret_str += f"Result {i+1}\n"
+            ret_str += f"Title: {d['title']}\n"
+            ret_str += f"{d['body']}\n"
+            ret_str += f"Source URL: {d['href']}\n"
+        return ret_str
+    else:
+        raise ValueError("Input must be dict or list[dict]")
